+++ lib/rrj/tools/tools.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'rrj/tools/log'
-require 'rrj/tools/env'
 require 'rrj/tools/config'
 require 'rrj/tools/requests'
 require 'rrj/tools/replaces/type'