+++ lib/rrj/errors/janus_response.rb
@@ -3,8 +3,11 @@
 module RubyRabbitmqJanus
   module Errors
     # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
+    #
     # Define an error primary for response
     class JanusResponse < JanusError
+      # Initialize a error for janus response module
+      # @param [String] message Text returning in raise
       def initialize(message)
         super "[Response] #{message}"
       end
@@ -13,31 +16,18 @@ module RubyRabbitmqJanus
     # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
     # Define an error for response initalize
     class JanusResponseInit < JanusResponse
+      # Initialize a error for janus response module in initializer
+      # @param [String] message Text returning in raise
       def initialize(message)
         super "Error create object : #{message}"
       end
     end
 
     # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
-    # Define an error for simple message
-    class JanusResponseSimple < JanusResponse
-      def initialize(message)
-        error = message['error']
-        super "[#{error['code']}] #{error['reason']} -- #{message}"
-      end
-    end
-
-    # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
-    # Define an error for plugin
-    class JanusResponsePlugin < JanusResponse
-      def initialize(message)
-        super "[#{message['error_code']}] #{message['error']}"
-      end
-    end
-
-    # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
     # Define an exception for json
     class JanusResponseJson < JanusResponse
+      # Initialize a error for janus response module in to_json
+      # @param [String] message Text returning in raise
       def initialize(message)
         super \
           "Error transform to JSON : #{message[0]} -- message : #{message[1]}"
@@ -47,6 +37,8 @@ module RubyRabbitmqJanus
     # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
     # Define an exception for nice_json
     class JanusResponsePrettyJson < JanusResponse
+      # Initialize a error for janus response module in to_nice_json
+      # @param [String] message Text returning in raise
       def initialize(message)
         super "Error transform to Pretty JSON : #{message}"
       end
@@ -55,6 +47,8 @@ module RubyRabbitmqJanus
     # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
     # Define an exception for hash
     class JanusResponseHash < JanusResponse
+      # Initialize a error for janus response module in to_hash
+      # @param [String] message Text returning in raise
       def initialize(message)
         super \
           "Error transform to Hash : #{message[0]} -- message : #{message[1]}"
@@ -62,16 +56,10 @@ module RubyRabbitmqJanus
     end
 
     # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
-    # Define an error for response plugin
-    class JanusResponsePluginData < JanusResponse
-      def initialize(message)
-        super "Error data response : #{message}"
-      end
-    end
-
-    # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
     # Define an error for response data
     class JanusResponseDataId < JanusResponse
+      # Initialize a error for janus response module in data_id
+      # @param [String] message Text returning in raise
       def initialize(message)
         super "Error Data : #{message}"
       end