+++ lib/rrj/errors/rabbit.rb
@@ -3,16 +3,10 @@
 module RubyRabbitmqJanus
   module Errors
     # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
-    # Define errors with Rabbitmq when request is not sending
-    class RequestNotExecuted < Errors::RRJError
-      def initialize(request)
-        super "Error in request executed : #{request}", :error
-      end
-    end
-
-    # @author VAILLANT Jeremy <jeremy.vaillant@dazzl.tv>
     # Define errors if connection to rabbitmq is failed
-    class ConnectionRabbitmqFailed < Errors::RRJError
+    class ConnectionRabbitmqFailed < RRJError
+      # Initialize a error for rabbit module. It's a fatal error
+      # @param [String] message Text returning in raise
       def initialize(message)
         super message, :fatal
       end