+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.1] - 2016-12-16
+### [Changed]
+- Fix handle reading if is sending to request
+- Delete branch repo for deploy action in travis
+
 ## [1.2.0] - 2016-12-15
 ### [Added]
 - dockerignore for travis